# MedicalAppointmentApp/MedicalAppointmentApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MedicalAppointmentApp.WebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Doctors": {
      "get": {
        "tags": [
          "Doctors"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorListItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorListItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorListItemDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Doctors"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/Doctors/{id}": {
      "get": {
        "tags": [
          "Doctors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Doctors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Doctors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Specializations": {
      "get": {
        "tags": [
          "Specializations"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecializationForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecializationForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecializationForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Specializations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecializationForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecializationForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SpecializationForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpecializationForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecializationForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecializationForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/Specializations/{id}": {
      "get": {
        "tags": [
          "Specializations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SpecializationForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecializationForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecializationForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Specializations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecializationForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecializationForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SpecializationForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Specializations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DoctorCreateDto": {
        "required": [
          "specializationId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "specializationId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DoctorForView": {
        "type": "object",
        "properties": {
          "doctorId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "specializationId": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "userAddressId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "specializationName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DoctorListItemDto": {
        "type": "object",
        "properties": {
          "doctorId": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "specializationName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DoctorUpdateDto": {
        "required": [
          "specializationId"
        ],
        "type": "object",
        "properties": {
          "specializationId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SpecializationForView": {
        "type": "object",
        "properties": {
          "specializationId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}